requirements.txt
Django==1.11.6
djangorestframework==3.7.1
drf-writable-nested==0.3.1
mysql-connector==2.1.4
mysqlclient==1.3.10
pytz==2017.2